+++ R/get_plot_forecast_data.R
@@ -22,7 +22,7 @@
 #' Default to `TRUE`.
 #' @param truth_source character specifying where the truth data will
 #' be loaded from if truth_data is not provided. Currently support `"JHU"`,
-#' `"USAFacts"`, `"NYTimes"` and `"HealthData"`.
+#' `NYTimes"` and `"HealthData"`.
 #' Optional if `truth_data` is provided.
 #' @param  target_variable_to_plot string specifying target type. It should be one of
 #' `"cum death"`, `"inc case"`, `"inc death"`, `"inc hosp"` and `"inc flu hosp"`.
@@ -61,7 +61,7 @@ get_plot_forecast_data <- function(forecast_data,
       "cum death", "inc case",
       "inc death", "inc hosp"
     )
-    valid_truth_sources <- c("JHU", "USAFacts", "NYTimes", "HealthData")
+    valid_truth_sources <- c("JHU", "NYTimes", "HealthData", "USAFacts")
   } else if (hub[1] == "ECDC") {
     valid_location_codes <- covidHubUtils::hub_locations_ecdc$location
     valid_target_variables <- c("inc case", "inc death")
@@ -122,6 +122,9 @@ get_plot_forecast_data <- function(forecast_data,
     }
   } else {
     # validate truth_source
+    if (tolower(truth_source) == "usafacts") {
+      stop("USAFacts can no longer be downloaded. Please use another truth source.")
+    }
     truth_source <- match.arg(truth_source,
       choices = valid_truth_sources,
       several.ok = FALSE